Remove Jest, use `downgrade-build` instead
The test run by Jest has been packaged up by extracting it from a few of
the packages I maintain.  So we don't need it here any more, and can
then get rid of Jest altogether.

Along with that, we may avoid loading Jest plugins for projects that
don't use Jest.  That makes it easier to test _this_ project, and any
others without Jest where we'd otherwise need to ignore the plugin's
rules.

Signed-off-by: Andrew Aylett <andrew@aylett.co.uk>